perf: Parallelize CI test execution and cut pipeline overhead
- jest maxWorkers is overridable via JEST_MAX_WORKERS; CI (local-test.sh)
  defaults it to min(cores, 8) since the suites are network-bound and the 30%
  default collapses to one worker on small runners
- split the serial stackmachine-sdk file into four files (deploy/apps/domains/
  wordpress) sharing tests/utils/stackmachine-sdk.ts helpers: 13 tests now run
  across workers, 890s -> 370s locally
- prepare/cache-warmup job exits early on an exact cache hit instead of
  booting the full stack (~9 min -> ~1 min on warm days; effective post-merge
  since callers pin the reusable workflow @main)
- WordPress suites now always test the latest WordPress release, resolved from
  api.wordpress.org per worker (currently 7.0, verified deploying locally);
  fallback to the previous pin only if the feed is unreachable
