+++ jest.config.js
@@ -2,8 +2,14 @@
 export default {
   globalSetup: "./jest-global-setup.ts",
   testEnvironment: "<rootDir>/tests/utils/preserve-failing-apps-environment.js",
-  // Max it! Most of the wait is in network block, nothing computationally heavy
-  maxWorkers: "30%",
+  // Most of the wait is network-bound, nothing computationally heavy. The
+  // percentage collapses to a single worker on small CI runners, so CI sets
+  // JEST_MAX_WORKERS explicitly (see local-platform/scripts/local-test.sh).
+  maxWorkers: process.env.JEST_MAX_WORKERS || "30%",
+  // Concurrency cap for test.concurrent within a single file (jest default 5).
+  // Overridable for experiments; raising it increases concurrent deploys and
+  // can overwhelm small runners hosting the co-resident local platform stack.
+  maxConcurrency: Number(process.env.JEST_MAX_CONCURRENCY || 5),
   silent: true,
   // Hard per-spec ceiling: no single test may run longer than 30 minutes. If it
   // does, jest fails it. Individual files may set a *stricter* limit via
